Move BaseOcrActivity from Activities folder to Contracts folder with namespace changed.
