+++ Cloud.Ocr/Cloud.Ocr.Activities/Activities/BankCardActivity.cs
@@ -3,6 +3,7 @@ using System.Activities;
 using System.Threading;
 using System.Threading.Tasks;
 using Cloud.Ocr.Activities.Properties;
+using Cloud.Ocr.Contracts;
 using Cloud.Ocr.Models;
 using UiPath.Shared.Activities;
 using UiPath.Shared.Activities.Localization;
