+++ Cloud.Ocr/Cloud.Ocr.Activities/Cloud.Ocr.Activities.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFrameworks>net461</TargetFrameworks>
 		<RootNamespace>Cloud.Ocr.Activities</RootNamespace>
