+++ Cloud.Ocr/Cloud.Ocr.Activities/Contracts/BaseOcrActivity.cs
@@ -1,4 +1,5 @@
-﻿using Cloud.Ocr.Activities.Properties;
+﻿using Cloud.Ocr.Activities;
+using Cloud.Ocr.Activities.Properties;
 using Cloud.Ocr.Contracts;
 using Cloud.Ocr.Models;
 using Newtonsoft.Json.Linq;
@@ -13,7 +14,7 @@ using UiPath.Shared.Activities;
 using UiPath.Shared.Activities.Localization;
 using UiPath.Shared.Activities.Utilities;
 
-namespace Cloud.Ocr.Activities
+namespace Cloud.Ocr.Contracts
 {
     public abstract class BaseOcrActivity : ContinuableAsyncCodeActivity
     {
